fix freedom wall
