+++ orgbee/angular.json
@@ -30,6 +30,7 @@
             "tsConfig": "tsconfig.app.json",
             "assets": ["src/favicon.ico", "src/assets"],
             "styles": [
+              "@angular/material/prebuilt-themes/indigo-pink.css",
               "src/styles.css",
               "./node_modules/@fortawesome/fontawesome-free/css/all.css"
             ],
@@ -84,6 +85,7 @@
             "tsConfig": "tsconfig.spec.json",
             "assets": ["src/favicon.ico", "src/assets"],
             "styles": [
+              "@angular/material/prebuilt-themes/indigo-pink.css",
               "src/styles.css",
               "./node_modules/@fortawesome/fontawesome-free/css/all.css"
             ],
