+++ orgbee/src/styles.css
@@ -30,3 +30,6 @@
 body {
   font-family: "Outfit", sans-serif;
 }
+
+html, body { height: 100%; }
+body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; }
